Update pywikibot from last stable release
A new "stable" tag was introduced which indicates a stable release of
pywikibot whereas master branch is under perpetual development and may
fail in some circumstanced during developing. This tag is updated
monthly when the tests passed or ealier if necessary e.g. upstream
changes.

The more stable release should be used for PAWS.

Bug: T217908