Rebuild Kubernetes, redesign the deployment, some fixes and upgrades
This involves a lot of changes to get things into a fully-understood,
secure and supported framework. This also includes several fixes that
have been needed or asked for for a long time that will be referenced
in the list of bugs. One that is not there is a bug in the way it mounts
dumps, which would never have been stable as it was.
Some of the commit messages are preserved below.

Bug: T211096
Bug: T188912
Bug: T167086
Bug: T258304
Bug: T157378
Bug: T195217

remove unused variable

add new oauth info

change baseurl to /

Fixing the PSP values

removed all root dependencies and fixed PSP and RBAC

set deployhook host to deploy-hook.paws.wmcloud.org as default for now

enable the ingress object generation

fix cloudmetadata entry

correct nodeport service mess

correct nodeport service mess again

remove auth for prometheus endpoint

switch to hub.paws.wmcloud.org

add paws-public to the same cluster -- if this works

Set the deployhook up to be optional for additional clusters

trying to figure out removing the path

switch the settings in public.yaml too

setting values to the current running image

add more flexibility to the deploy-hook to ease transition

Proposed legacy URL ingress

Fixing the labels on the public and legacy ingresses

switch to a simple redirect instead of snippit

removed the * from the capture

critical fix of the template value for the paws legacy host

correct the values path for the jupyterhub ingress

Add a redirect to wikitech for the front page /

URL encode path to file after replacing user name with user id

Make URLs safe for processing by nginx before returning from LUA by reapplying URL encoding following replacement of the `User:<name>` path component with the associated Wikimedia SUL account's uid.

Bug: T258304

Revert "URL encode path to file after replacing user name with user id"

URL encode path to file after replacing user name with user id

Make URLs safe for processing by nginx before returning from LUA by
reapplying URL encoding following replacement of the `User:<name>` path
component with the associated Wikimedia SUL account's uid. See
<https://github.com/openresty/lua-nginx-module#ngxescape_uri> for the
docs on using `ngx.escape_uri`. We need the escape to run as a full URI
escape in this case so that whitespace is encoded, but path separators
are not.

Bug: T258304

Revert "URL encode path to file after replacing user name with user id"

nbserve: handle file paths containing whitespace

Nginx wants paths passed to its rewrite directive to be single quoted
(binary) if they contain whitespace. The rewrite_by_lua equivalent is
achieved by using the third argument to `ngx.req.set_uri` which was
introduced in lua-nginx-module at 5d15172 (v0.10.16). We need openresty
at 11fbbe5 (v1.17.8.1) or newer to pick up this functionality.
v1.17.8.2 is the latest release at the time this patch was written.

Bug: T258304

update the location of the nbserve image

set defaults that work for the rebuilt cluster for deploy-hook and db-proxy images

set the newbuild tags as defaults for the rebuilt cluster on the other images

switch the default tag to latest for all images

Update the pawspublic link to be consistent with new cluster

Fix up the pull processing for the singleuser image a bit

Fix capital letter in pull policy

Update the paws hacks a bit

Co-authored-by: Bryan Davis <bd808@wikimedia.org>
Co-authored-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>