ci: Adding linting steps to travis
As a start for getting CI working, add tox and linting. Also, disable
the deployhook step on merge to master until a couple bugs are worked out.

As is the deployhook misses some image pulls, has the wrong URL and
may or may not actually work just yet in the new cluster
(testing coming soon!) I am sure that it will deploy the wrong images
without modification, if it succeeds in deploying images.

Bug: T256689