multiinstance-replicas: Select host backend based on database requested
This requires a change in how users behave, but it is a small-ish change.
Users just need to be explicit about which database they are connecting to
in order to get to the right place (if this works on an actual cluster).

The template renders correctly (tested with `helm template ..`), but this
change to the lua:

1. Must read a default database from the connection strings, which seems likely.
2. Can quickly select the right section alias for the largest popular wikis:
enwiki, wikidata and commonswiki.
3. Has a much slower process to find the section alias for other wikis, involving
web requests to each section's dblist from noc.wikimedia.org until it finds
the right one.

That last step seems necessary (or something like it) in order to choose
a backend that is declared *at the start of the proxy process*. We could just
count on the aliased names in DNS otherwise.

Bug: T260389