developers: create a a dev environment and fix the db-proxy with it
This commit moves some "default" values that are aimed only at production
into the secrets.yaml file (which overrides values.yaml on install). That
allows values.yaml to contain dummy values and defaults that allow installing
disposable dev environments from helm without modification.

This adds a NOTES.txt file, which is templated out and is displayed on
"helm install" so that specific commands and values can be shown to the
user to complete their dev environment.

Since Bitnami has the only helm repository with a public and functioning
mediawiki, we are using that for mediawiki and mariadb for now. They are
widely used charts that are well-regarded and professionally maintained.

Bug: T260389