remove openrefine pending security issue resolution
Openrefine creates a workspace.json with cleartext secrets in
it that seems to just do whatever the umask says. Since PAWS is public
by default, that means the information is on the web.

This is actually already applied everywhere because the software also
changes the file back if you try to chmod it, meaning that it being
present is enough to keep opening your file to the internet.

Bug: T283839