hub database: move from ToolsDB to Trove
Trove has been deployed for a bit. Since we don't have replication or
backups for it in WMCS yet, I've set up an instance that backs up
mysql dumps to NFS. Honestly, the data is likely far safer there than
in ToolsDB in my opinion.

Please note, this is already deployed to settle a security issue.

Bug: T289180