Install a specific version of Pywikibot
Installing the "stable" branch has the benefit of using the latest
release whenever the image is rebuilt. However, that means to get the
latest Pywikibot version, someone will need to do a no-change rebuild of
the image, bypassing Git and the standard CI/CD processes.

Specifying an explicit version means this will just work with docker
caching and cause a commit to go through CI/CD like all other changes.
It does mean that a new commit needs to be made for each new version.
LibUp will file a task in the #PAWS Phabricator project[1] every time a
new Pywikibot version is released so it doesn't get forgotten about.

I put the PYWIKIBOT_VERSION env at the top for easier readability, which
means bumping Pywikibot will invalidate the whole build cache. I'm not
sure if that's a good thing.

[1] https://gerrit.wikimedia.org/r/723018