minikube on 1.23 k8s 1.20.11
NOTE! This is broken! jupyter appears to be replacing
capital letters with -<hex> so T becomes -54, which makes a
problem for one of the labels that it tries to make as it
starts any username that begins with a capital letter (so
almost all of them) with a - which is forbidden. This will
need to be resolved before this can be used. We get around
this for testing by putting a number in front of the admin
user.

This updates the jupyter and mediawiki versions allowing for
paws to be run on a modern minikube, as the newer ingress
wasn't allowing the old ingress api.

the singleuser and paws-hub containers were rebuilt using
this code and will need to be updated in the registry for
this to work on a wider scale.