Move to github actions from Travis
This change will remove travis from the workflow. Using github actions in its place. Additionally image creation will be preformed on a PR that has changes in an associated image directory, such as images/singleuser. If there are changes as part of the PR in such a directory the image will be rebuilt, and pushed to quay using the pr number as its tag, such as "pr-23", and then paws/values.yaml will be updated to reflect this. This allows us to automatically create new tags and update the deployment to match. Someone interested in testing the deploy on minikube should only need to pull the branch and run a helm install/upgrade, where previously there would have been manual building of the container, and editing values.yaml or an override file. Additionally this gets us using the same container we tested in prod. Finally this brings us closer to being able to automate deploys. By using a unique tag on each container we should only need to run the helm upgrade command on a prod system to be upgraded with any container or configuration changes. Such can be added as another action or webhook that is triggered when a PR is merged.

This will need secrets before it can work. We will need to decide what creds to put in QUAY_USER and QUAY_PASSWORD in the github repo secrets.

In the lint.yaml in PR 70 there was an additional test for the kubeval:
./kubeval/kubeval -v $KUBERNETES_VERSION --strict ingress/* manifests/*
which in travis was:
kubeval -v $K8S_VERSION --strict manifests/*

ingress was failing so leaving it off for now. T295230 is a ticket to update this.

toolforge/paws/.github/workflows/build-and-push.yaml@T292442
Should either be moved to its own repo. Or we should apply a tag referencing that there is a workflow here. Either way we will have to update it when we merge.