Move nbserve to upstream container
The current nbserve container is using a base of
docker-registry.tools.wmflabs.org/toolforge-python37-sssd-base:latest
this is likely an artifact from when PAWS was a toolforge
project. Moving to an ubuntu container both helps to
decouple PAWS from particular environments, and helps to
standardize PAWS containers to one another other (Ubuntu
being the popular image). As a bonus this image is about
half the size of the previous image (566MB vs 1.15GB)

Bug: T308926