construct nginx with config map (#227)
* construct nginx with config map

This change removes a startup script that constructs
nginx.conf via python then launches openresty when it
completes. The python was used to determine the nameserver
and the address of renderer. The former can be removed by
using openresty's "local=on" param and the latter can be
managed using k8s dns.

Bug: T322205

* auto update of tag

Co-authored-by: Github Action <auto@github.com>